+++ src/patterns/rainbow.rs
@@ -4,13 +4,13 @@ use smart_leds::RGB8;
 
 pub const DURATION: Duration = Duration::from_millis(10);
 
-pub fn generate(data: &mut [RGB8], counter: &mut u16) {
+pub fn generate(data: &mut [RGB8], counter: &mut u8) {
     // Advance the animation counter
     *counter = counter.wrapping_add(1);
     let j = *counter;
 
     for i in 0..data.len() {
-        let wheel_pos = ((i as u16 * (256 / data.len() as u16) + j) & 255) as u8;
+        let wheel_pos = (((i * 256) as u16 / data.len() as u16 + j as u16) & 255) as u8;
         data[i] = wheel(wheel_pos);
     }
 }
